chore：捐贈下單

+++ lib/models/charity_favorites_item_detail.dart
@@ -4,9 +4,9 @@ class CharityFavoritesItemDetailVo {
   int? orderStatus;
   String? bussinessName;
   String? bussinessNo;
-  int? buyPrice;
+  double? buyPrice;
   String? originalPrice;
-  int? buyPriceForCharity;
+  double? buyPriceForCharity;
   int? forCharityMin;
   int? charityId;
   String? charityImg;
